// src/app/core/models/user.ts
export interface User {
    firstName: string;
    lastName: string;
    phone: string;
    degree: string;
    employment: string;
    income: number;
    savings: number;
    debt: number;
    age: number;
    zipcode: string;
    haveVehicle: boolean;
    haveHouse: boolean;
    havePet: boolean;
    have_health_cover: string;
    have_vision_cover: string;
    have_dental_cover: string;
    have_life_cover: string;
    have_longtermdisability_cover: string;
    have_shorttermdisability_cover: string;
    have_accident_cover: string;
    have_criticalillness_cover: string;
    have_auto_cover: string;
    have_homeowner_cover: string;
    have_renters_cover: string;
    have_pet_cover: string;
}